adjust NetAddress implementation and Protocol

+++ p2p/msg/pong.go
@@ -6,9 +6,9 @@ type Pong struct {
 	Ping
 }
 
-func NewPong(nonce uint32) *Pong {
+func NewPong(nonce uint64) *Pong {
 	pong := new(Pong)
-	pong.Nonce = uint64(nonce)
+	pong.Nonce = nonce
 	return pong
 }
